'deleteconfirm'

+++ src/views/picture_article/comment-list.vue
@@ -97,14 +97,23 @@ export default {
       this.getComment()
     },
     deleteData(row) {
-      const delData = { id: row.id }
-      deleteComment(delData).then(() => {
-        this.getComment()
-        this.$notify({
-          title: '成功',
-          message: '删除成功',
-          type: 'success',
-          duration: 2000
+      this.$confirm('此操作将永久删除该文件, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        const delData = { id: row.id }
+        deleteComment(delData).then(() => {
+          this.getComment()
+          this.$message({
+            type: 'success',
+            message: '删除成功!'
+          })
+        })
+      }).catch(() => {
+        this.$message({
+          type: 'info',
+          message: '已取消删除'
         })
       })
     }
